Fully implemented caching. Will test later, since it's quite late now.

+++ src/main/java/com/example/argusclone/services/CourseDeletionService.java
@@ -1,5 +1,10 @@
 package com.example.argusclone.services;
 
+import com.example.argusclone.entities.Course;
+import com.example.argusclone.exceptions.OperationNotAllowedYetException;
+import com.example.argusclone.exceptions.ResourceNotFoundException;
+import com.example.argusclone.repositories.CourseRepository;
+import com.example.argusclone.services.caching.CourseRelatedCacheService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -7,22 +12,33 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class CourseDeletionService {
     private final CourseService courseService;
-    private final LectureService lectureService;
-    private final ScoreService scoreService;
+    private final CourseRepository courseRepository;
+    private final CourseRelatedCacheService courseRelatedCacheService;
 
     @Autowired
     public CourseDeletionService(CourseService courseService,
-                                 LectureService lectureService,
-                                 ScoreService scoreService) {
+                                 CourseRepository courseRepository,
+                                 CourseRelatedCacheService courseRelatedCacheService) {
         this.courseService = courseService;
-        this.lectureService = lectureService;
-        this.scoreService = scoreService;
+        this.courseRepository = courseRepository;
+        this.courseRelatedCacheService = courseRelatedCacheService;
     }
 
     @Transactional
     public void deleteCourseById(Integer id) {
-        scoreService.deleteScoresByCourseId(id);
-        lectureService.deleteLecturesByCourseId(id);
+        Course course = courseRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Course with an id of " + id + " not found")
+        );
+
+        if (course.getGroups() != null || !course.getGroups().isEmpty()) {
+            throw new OperationNotAllowedYetException("Course can't be deleted because it has active groups. Delete them first.");
+        }
+
+        if (course.getScores() != null || !course.getScores().isEmpty()) {
+            throw new OperationNotAllowedYetException("Course can't be deleted because it has active scores. Delete them first.");
+        }
+
+        courseRelatedCacheService.clearAllRelevantCachesForCourse(id);
         courseService.deleteCourseById(id);
     }
 }
